+++ CHANGELOG.md
@@ -53,6 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When calculating the total wear time during waking hours, waking hours were not considered correctly and were affected
   by the daylight saving state at the time the script was run! This is now fixed, by not using actual dates for the
   comparison, but just seconds from midnight.
+- `apply_transformation` now returns a normal index instead of MultiIndex, when all expected output col names are 
+  strings (https://github.com/mobilise-d/mobgap/pull/199)
 
 ### Removed